Changes for the 1.15.1 release
Adds the ability to release from a branch other than "develop"
to mkrelease.  While testing, there was some trouble with
the htscodecs submodule not updating correctly, so this is now
done explicitly.

Makes tag_release start git mergetool if the merge fails.  This
is most likely due to changes in NEWS between develop and the
released branch, and can usually be fixed easily allowing the
script to continue.