Changes for samtools-1.16.1; allow differing package versions.
Updates the release infrastructure so that it's possible to
have differing version numbers for the various packages, and
it's easier to make point releases for single packages (in this
case samtools).

Settings like the version to release are moved to a new file
so they can be shared between the various scripts and (rather
hackily) the Makefile.  It's possible to set different branches
and version numbers for individual packages by setting, e.g.,
`samtools_version`, `samtools_branch` variables to override the
default `version` and `branch` ones.

It's not necessary to have all of samtools, bcftools and htslib
in the `packages` variable.  If htslib is left out, the master
branch will be selected so the latest release is available for
building the samtools or bcftools tar files.  The Makefile is
adjusted so that it only builds the requested packages (plus
htslib, if needed for the others).